+++ tests/app/src/Bootloader/AuthBootloader.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Spiral\App\Bootloader;
 
 use Spiral\Boot\Bootloader\Bootloader;
-use Spiral\Boot\BootloadManager\CustomizableBootloadManager;
+use Spiral\Boot\BootloadManager\StrategyBasedBootloadManager;
 use Spiral\Bootloader\Auth\HttpAuthBootloader;
 
 class AuthBootloader extends Bootloader
 {
-    public function init(CustomizableBootloadManager $bootloadManager): void
+    public function init(StrategyBasedBootloadManager $bootloadManager): void
     {
         $bootloadManager->bootload([HttpAuthBootloader::class]);
     }